Add files via upload
Do you know what a palindrome is?
For those of you who have not heard of it before, a palindrome is an expression that runs both ways. In other words, it is spelled the same front and backwards. “A man, a plan, a canal—Panama”, “Never Odd Or Even” or “Madam In Eden, I’m Adam”, are some of many examples.
I created a function that identifies if an expression is a palindrome or not. Here is the project description:
- Language: Python 3
- IDE: Jupyter Notebook